refactor(react-components-storybook): colocate ObjectTable stories in Features files
Replace the large objectTableStoryDefs.tsx re-export module with a small
objectTableStoryHelpers.tsx that holds only the shared pieces (the argTypes
meta, the shared column definitions, and the interaction-test helpers). Each
story is now defined directly in the Features/*.stories.tsx file where it is
displayed (and Default in the component-root file), importing the shared
helpers instead of re-exporting from a monolith. No story content changes.
